feat: update patch version to 1.0.9 and add sanitizeRichTextForSetText function for rich text validation
