+++ public/manifest.json
@@ -10,7 +10,7 @@
   "version": {
     "major": 1,
     "minor": 0,
-    "patch": 8
+    "patch": 9
   },
   "unlisted": false,
   "theme": [],
